+++ obj/src/connect/MongoDbConnectionResolver.js
@@ -109,20 +109,12 @@ class MongoDbConnectionResolver {
                 hosts += ',';
             hosts += host + (port == null ? '' : ':' + port);
         }
-        // Define database
-        let database = '';
-        for (let connection of connections) {
-            database = database || connection.getAsNullableString("database");
-        }
-        if (database.length > 0) {
-            database = '/' + database;
-        }
         // Define authentication part
         let auth = '';
         if (credential) {
-            let username = credential.getUsername();
+            let username = encodeURIComponent(credential.getUsername());
             if (username) {
-                let password = credential.getPassword();
+                let password = encodeURIComponent(credential.getPassword());
                 if (password) {
                     auth = username + ':' + password + '@';
                 }
@@ -155,7 +147,7 @@ class MongoDbConnectionResolver {
             params = '?' + params;
         }
         // Compose uri
-        let uri = "mongodb://" + auth + hosts + database + params;
+        let uri = "mongodb://" + auth + hosts + params;
         return uri;
     }
     /**