+++ test/persistence/DummyMongoDbPersistence.test.ts
@@ -12,6 +12,8 @@ suite('DummyMongoDbPersistence', ()=> {
     let mongoHost = process.env['MONGO_SERVICE_HOST'] || 'localhost';
     let mongoPort = process.env['MONGO_SERVICE_PORT'] || 27017;
     let mongoDatabase = process.env['MONGO_DB'] || 'test';
+    let mongoUser = process.env['MONGO_USER'] || '';
+    let mongoPass = process.env['MONGO_PASS'] || '';
 
     if (mongoUri == null && mongoHost == null) {
         return;
@@ -22,7 +24,9 @@ suite('DummyMongoDbPersistence', ()=> {
             'connection.uri', mongoUri,
             'connection.host', mongoHost,
             'connection.port', mongoPort,
-            'connection.database', mongoDatabase
+            'connection.database', mongoDatabase,
+            'credential.username', mongoUser,
+            'credential.password', mongoPass,
         );
 
         persistence = new DummyMongoDbPersistence();