+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> MongoDB components for Pip.Services in Node.js / ES2017 Changelog
 
+## <a name="1.1.1"></a> 1.1.1 (2023-03-28) 
+
+### Bug fixes
+* Fixed unescaped user and password bug
+* Fixed connection string
+
 ## <a name="1.1.0"></a> 1.1.0 (2023-02-07) 
 
 ### Features